refactor: [las] split model weight impl to seperate file
